Avoid re-testing values which are known in advance to be bad.
The original algorithm would always bump the value up twice as much as it cut it down the last time, even though it's obvious in advance that that value will be bad.

Now, once the equation error rolls over zero, calculator will switch to another algorithm and keep splitting the interval between two bad values in half until a satisfactory value is found.

This significantly reduces the amount of iterations required to find an APR value that is precise enough.